+++ common/unsafe/src/main/java/org/apache/spark/unsafe/array/ByteArrayMethods.java
@@ -43,13 +43,9 @@ public class ByteArrayMethods {
   // Some JVMs can't allocate arrays of length Integer.MAX_VALUE; actual max is somewhat smaller.
   // Be conservative and lower the cap a little.
   // Refer to "http://hg.openjdk.java.net/jdk8/jdk8/jdk/file/tip/src/share/classes/java/util/ArrayList.java#l229"
-  public static int MAX_ARRAY_LENGTH = Integer.MAX_VALUE - 8;
-
-  // Use this value if the allocated byte arrays are used to store other types rather than bytes.
-  // This is equivalent to `Integer.MAX_VALUE - 15`
-  public static int maxWordRoundedArrayLength() {
-    return roundNumberOfBytesToNearestWord(MAX_ARRAY_LENGTH - 8);
-  }
+  // This value is word rounded. Use this value if the allocated byte arrays are used to store other
+  // types rather than bytes.
+  public static int MAX_ROUNDED_ARRAY_LENGTH = Integer.MAX_VALUE - 15;
 
   private static final boolean unaligned = Platform.unaligned();
   /**
